// src/backend/otel/proto/opentelemetry/proto/common/v1/common.ts
export interface AnyValue {
  /**
   * @generated from protobuf oneof: value
   */
  value: {
    oneofKind: "stringValue";
    /**
     * @generated from protobuf field: string string_value = 1
     */
    stringValue: string;
  } | {
    oneofKind: "boolValue";
    /**
     * @generated from protobuf field: bool bool_value = 2
     */
    boolValue: boolean;
  } | {
    oneofKind: "intValue";
    /**
     * @generated from protobuf field: int64 int_value = 3
     */
    intValue: bigint;
  } | {
    oneofKind: "doubleValue";
    /**
     * @generated from protobuf field: double double_value = 4
     */
    doubleValue: number;
  } | {
    oneofKind: "arrayValue";
    /**
     * @generated from protobuf field: opentelemetry.proto.common.v1.ArrayValue array_value = 5
     */
    arrayValue: ArrayValue;
  } | {
    oneofKind: "kvlistValue";
    /**
     * @generated from protobuf field: opentelemetry.proto.common.v1.KeyValueList kvlist_value = 6
     */
    kvlistValue: KeyValueList;
  } | {
    oneofKind: "bytesValue";
    /**
     * @generated from protobuf field: bytes bytes_value = 7
     */
    bytesValue: Uint8Array;
  } | {
    oneofKind: undefined;
  };
}
/**
 * @generated from protobuf message opentelemetry.proto.common.v1.ArrayValue
 */
export interface ArrayValue {
  /**
   * @generated from protobuf field: repeated opentelemetry.proto.common.v1.AnyValue values = 1
   */
  values: AnyValue[];
}
/**
 * @generated from protobuf message opentelemetry.proto.common.v1.KeyValueList
 */
export interface KeyValueList {
  /**
   * @generated from protobuf field: repeated opentelemetry.proto.common.v1.KeyValue values = 1
   */
  values: KeyValue[];
}
/**
 * @generated from protobuf message opentelemetry.proto.common.v1.KeyValue
 */
export interface KeyValue {
  /**
   * @generated from protobuf field: string key = 1
   */
  key: string;
  /**
   * @generated from protobuf field: opentelemetry.proto.common.v1.AnyValue value = 2
   */
  value?: AnyValue;
}
// ...
export const AnyValue = new AnyValue$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ArrayValue$Type extends MessageType<ArrayValue> {
  constructor() {
    super("opentelemetry.proto.common.v1.ArrayValue", [
      { no: 1, name: "values", kind: "message", repeat: 2, /*RepeatType.UNPACKED*/ T: () => AnyValue },
    ]);
  }
  override create(value?: PartialMessage<ArrayValue>): ArrayValue {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.values = [];
    if (value !== undefined) {
      reflectionMergePartial<ArrayValue>(this, message, value);
    }
    return message;
  }
  override internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: ArrayValue,
  ): ArrayValue {
    let message = target ?? this.create(), end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated opentelemetry.proto.common.v1.AnyValue values */ 1:
          message.values.push(AnyValue.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw") {
            throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          }
          let d = reader.skip(wireType);
          if (u !== false) {
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
          }
      }
    }
    return message;
  }
  override internalBinaryWrite(message: ArrayValue, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* repeated opentelemetry.proto.common.v1.AnyValue values = 1; */
    for (let i = 0; i < message.values.length; i++) {
      AnyValue.internalBinaryWrite(message.values[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
    }
    let u = options.writeUnknownFields;
    if (u !== false) {
      (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    }
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.common.v1.ArrayValue
 */
export const ArrayValue = new ArrayValue$Type();
// @generated message type with reflection information, may provide speed optimized methods
class KeyValueList$Type extends MessageType<KeyValueList> {
  constructor() {
    super("opentelemetry.proto.common.v1.KeyValueList", [
      { no: 1, name: "values", kind: "message", repeat: 2, /*RepeatType.UNPACKED*/ T: () => KeyValue },
    ]);
  }
  override create(value?: PartialMessage<KeyValueList>): KeyValueList {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.values = [];
    if (value !== undefined) {
      reflectionMergePartial<KeyValueList>(this, message, value);
    }
    return message;
  }
  override internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: KeyValueList,
  ): KeyValueList {
    let message = target ?? this.create(), end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated opentelemetry.proto.common.v1.KeyValue values */ 1:
          message.values.push(KeyValue.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw") {
            throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          }
          let d = reader.skip(wireType);
          if (u !== false) {
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
          }
      }
    }
    return message;
  }
  override internalBinaryWrite(
    message: KeyValueList,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* repeated opentelemetry.proto.common.v1.KeyValue values = 1; */
    for (let i = 0; i < message.values.length; i++) {
      KeyValue.internalBinaryWrite(message.values[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
    }
    let u = options.writeUnknownFields;
    if (u !== false) {
      (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    }
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.common.v1.KeyValueList
 */
export const KeyValueList = new KeyValueList$Type();
// @generated message type with reflection information, may provide speed optimized methods
class KeyValue$Type extends MessageType<KeyValue> {
  constructor() {
    super("opentelemetry.proto.common.v1.KeyValue", [
      { no: 1, name: "key", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: "value", kind: "message", T: () => AnyValue },
    ]);
  }
  override create(value?: PartialMessage<KeyValue>): KeyValue {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.key = "";
    if (value !== undefined) {
      reflectionMergePartial<KeyValue>(this, message, value);
    }
    return message;
  }
  override internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: KeyValue,
  ): KeyValue {
    let message = target ?? this.create(), end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string key */ 1:
          message.key = reader.string();
          break;
        case /* opentelemetry.proto.common.v1.AnyValue value */ 2:
          message.value = AnyValue.internalBinaryRead(reader, reader.uint32(), options, message.value);
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw") {
            throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          }
          let d = reader.skip(wireType);
          if (u !== false) {
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
          }
      }
    }
    return message;
  }
  override internalBinaryWrite(message: KeyValue, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* string key = 1; */
    if (message.key !== "") {
      writer.tag(1, WireType.LengthDelimited).string(message.key);
    }
    /* opentelemetry.proto.common.v1.AnyValue value = 2; */
    if (message.value) {
      AnyValue.internalBinaryWrite(message.value, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
    }
    let u = options.writeUnknownFields;
    if (u !== false) {
      (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    }
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.common.v1.KeyValue
 */
export const KeyValue = new KeyValue$Type();